refactor(server): extract music + achievements helpers from Game into mixins
Move the music-mood/track methods and the end-of-match medal tally out of the
Game class into music.js and achievements.js, mixed back in via
Object.assign(Game.prototype, ...). Methods keep `this` and all call sites are
unchanged. game.js drops to ~2285 lines.

Pure mechanical move, no logic change; smoke test passes on all 50 maps.
