+++ server/game.js
@@ -599,83 +599,6 @@ class Game {
 		messenger.messageRoomBySig(this.roomSig, "startWaiting", null);
 		this.currentState = this.stateMap.waiting;
 	}
-	//The mood the room should be hearing right now: exciting overrides everything
-	//when anyone is one notch from winning, otherwise brutal rounds get brutal music.
-	computeMusicMood() {
-		for (var id in this.playerList) {
-			if (this.playerList[id].nearVictory == true) {
-				return "exciting";
-			}
-		}
-		if (this.gameBoard.brutalRound == true) {
-			return "brutal";
-		}
-		return "calm";
-	}
-	//Pick a track name for a mood from the shared config manifest, avoiding an
-	//immediate repeat of avoidTrack when the playlist has alternatives.
-	pickMusicTrack(mood, avoidTrack) {
-		var playlist = c.music[mood];
-		if (playlist == null || playlist.length == 0) {
-			return null;
-		}
-		if (playlist.length == 1) {
-			return playlist[0];
-		}
-		var track = playlist[utils.getRandomInt(0, playlist.length - 1)];
-		var attempts = 0;
-		while (track == avoidTrack && attempts < 5) {
-			track = playlist[utils.getRandomInt(0, playlist.length - 1)];
-			attempts++;
-		}
-		return track;
-	}
-	//Set the room's music and remember when, so the fallback timer can recover
-	//if no client ever reports the track ending.
-	setRoomMusic(mood, track) {
-		this.currentMusic = { mood: mood, track: track };
-		this.musicChangedAt = Date.now();
-	}
-	//Called every racing tick: if the room's mood changed (someone hit/left
-	//near-victory), pick a track for the new mood and tell every client to switch.
-	updateMusicMood() {
-		if (this.currentMusic == null) {
-			return;
-		}
-		var desiredMood = this.computeMusicMood();
-		if (desiredMood == this.currentMusic.mood) {
-			return;
-		}
-		this.setRoomMusic(desiredMood, this.pickMusicTrack(desiredMood, null));
-		messenger.messageRoomBySig(this.roomSig, "musicMood", this.currentMusic);
-	}
-	//A client reported its background track finished. Background tracks don't loop,
-	//so pick the next one for the current mood and broadcast it, keeping music
-	//continuous and in sync. Stale reports for an already-rotated track are ignored.
-	rotateMusicTrack(endedTrack) {
-		if (this.currentMusic == null || this.currentMusic.track != endedTrack) {
-			return;
-		}
-		var mood = this.currentMusic.mood;
-		this.setRoomMusic(mood, this.pickMusicTrack(mood, endedTrack));
-		messenger.messageRoomBySig(this.roomSig, "musicMood", this.currentMusic);
-	}
-	//Safety net: clients normally drive rotation by reporting "ended" (precise,
-	//per-track). But if every client is muted/backgrounded/autoplay-blocked, that
-	//report never comes. If the current track has been playing past the fallback
-	//window (set above the longest track so it never cuts normal playback), advance
-	//it anyway so the room never gets stuck in silence.
-	checkMusicFallback() {
-		if (this.currentMusic == null || this.musicChangedAt == null) {
-			return;
-		}
-		if (Date.now() - this.musicChangedAt < c.musicFallbackSeconds * 1000) {
-			return;
-		}
-		var mood = this.currentMusic.mood;
-		this.setRoomMusic(mood, this.pickMusicTrack(mood, this.currentMusic.track));
-		messenger.messageRoomBySig(this.roomSig, "musicMood", this.currentMusic);
-	}
 	startLobby() {
 		console.log("Start Lobby")
 		debug.log("startLobby: from state=", this.currentState, " playerCount=", this.playerCount);
@@ -842,75 +765,8 @@ class Game {
 		this.currentState = this.stateMap.gameOver;
 		messenger.messageRoomBySig(this.roomSig, 'startGameover', { winner: player, achievements: this.gatherAchievements() });
 	}
-	gatherAchievements() {
-		var achievements = {
-			mostKills: { ids: [], value: 0, title: "Serial killer" },
-			savior: { ids: [], value: 0, title: "Savior" },
-			survivalist: { ids: [], value: 0, title: "Survivalist" },
-			brutalist: { ids: [], value: 0, title: "Brutalist" },
-			mostMurdered: { ids: [], value: 0, title: "Picked on" },
-			resourceful: { ids: [], value: 0, title: "Resouceful" },
-			bully: { ids: [], value: 0, title: "Bully" },
-			doubleKill: { ids: [], value: 0, title: "Double Kill" },
-			tripleKill: { ids: [], value: 0, title: "Triple Kill" },
-			megaKill: { ids: [], value: 0, title: "Mega Kill" },
-		};
-		for (var id in this.playerList) {
-			var player = this.playerList[id];
-
-			//Best Murderer
-			this.checkForNewMedalHolder(achievements.mostKills, id, player.totalKills);
-
-			//Savior
-			if (player.savior > 0) {
-				achievements.savior.ids.push(id);
-			}
-
-			//Survivalist
-			this.checkForNewMedalHolder(achievements.survivalist, id, player.survivalist);
-
-			//Brutalist
-			this.checkForNewMedalHolder(achievements.brutalist, id, player.brutalist);
-
-			//Bully
-			this.checkForNewMedalHolder(achievements.bully, id, player.bully);
-
-			//Resourceful
-			this.checkForNewMedalHolder(achievements.resourceful, id, player.resourceful);
-
-			//Picked on
-			var mostKilled = null;
-			for (var i = 0; i < player.killedPlayerList.length; i++) {
-				var murderedID = player.killedPlayerList[i];
-				if (mostKilled == null) {
-					mostKilled = {};
-					mostKilled[murderedID] = 1;
-				} else {
-					mostKilled[murderedID] += 1;
-				}
-			}
-			if (mostKilled != null) {
-				for (var murderID in mostKilled) {
-					this.checkForNewMedalHolder(achievements.mostMurdered, murderID, mostKilled[murderID]);
-				}
-			}
-		}
-		return achievements;
-	}
-	checkForNewMedalHolder(medal, id, value) {
-		if (value == 0) {
-			return;
-		}
-		if (value >= medal.value) {
-			if (value > medal.value) {
-				medal.ids = [id];
-			} else {
-				medal.ids.push(id);
-			}
-			medal.value = value;
-		}
-	}
 }
+Object.assign(Game.prototype, require('./music.js'), require('./achievements.js'));
 
 
 class GameBoard {
